requirements.txt
flask==2.3.2
sentence-transformers==2.2.2
faiss-cpu==1.7.4
numpy==1.24.4
scikit-learn==1.3.0
pillow==10.0.0
pdf2image==1.16.3
pytesseract==0.3.10
tqdm==4.65.0
google-generativeai==0.6.1
